outlier-detection/requirements.txt
numpy==1.17.0               
opencv-python==4.2.0.34               
pillow==7.1.0                    
pydicom==1.1.0               
python==3.7.6                
torch==1.5.0+cu101             
torchvision==0.6.0+cu101
